[ADD]执行器调用接口从多线程调用改为rabbitmq生产消费

+++ xtimer/src/main/java/com/rice/lfcdemo/entity/TaskModel.java
@@ -16,7 +16,7 @@ public class TaskModel extends BaseModel implements Serializable {
 
     private Long runTimer;
 
-    private int costTime;
+    private Long costTime;
 
     private int status;
 
@@ -60,11 +60,11 @@ public class TaskModel extends BaseModel implements Serializable {
         this.runTimer = runTimer;
     }
 
-    public int getCostTime() {
+    public Long getCostTime() {
         return costTime;
     }
 
-    public void setCostTime(int costTime) {
+    public void setCostTime(Long costTime) {
         this.costTime = costTime;
     }
